requirements.txt
django-cms==3.11.5
flake8
